Rewrite cpufreqctl script to support several backends
This commit adds a modular architecture to the cpufreqctl script which supports
multiple backends (e.g. intel_pstate) whereas one is a fake backend. This is in
preparation for #145 and future updates of the extension to support easy testing
in virtual machine environments. The backend is selected over an environment
variable. This will also make the integration of AMD Ryzen support possible, if
not directly over zenstates, it will be possible over the cpufreq
scaling-driver, see #148.

Missing points:
 - integration of new cli into extension code
 - min/max frequency check
 - for some cpufreq scaling-drivers, scaling_available_frequencies must be
   respected, no min/max values available but complete discrete steps