Fix dismiss button requiring 2 clicks
The dismiss and save button are now only made sensitive when the values
of the controls actually changed. Previously, any change event of the
controls would lead to sensitive discard and save buttons. Now as lower
and upper boundaries are set to the minimum and maximum frequency scales
the controls might trigger change events even if the actual value has
not changed. By checking for difference with the actual profile values
this can be avoided.